Del opp regnskaps- og hjelpefunksjoner
