+++ nordlys/regnskap/__init__.py
@@ -1,0 +1,11 @@
+"""Offentlig API for regnskapsanalyse."""
+
+from .analysis import AnalysisRow, compute_balance_analysis, compute_result_analysis
+from .prep import prepare_regnskap_dataframe
+
+__all__ = [
+    "AnalysisRow",
+    "prepare_regnskap_dataframe",
+    "compute_balance_analysis",
+    "compute_result_analysis",
+]
